feat: Ensure puzzle is solveable
Prior to this change, it was possible to generate unsolveable puzzles.
This change fixes that, by:

 1. Generate a grid
 2. Pick a random word that's bigger than the grid
 3. Stuff that word into the grid
    - This ensures that every puzzle _can_ be solved, and with one word
 4. Present the grid

The solver will then re-solve the puzzle, potentially finding a shorter
solution than the one inserted -- but definitely at least finding the
synthetic spangram.

This also changes from an A* search to a BFS search. This simplified the
algorithm, but also improved its performance drastically.