fix: Only break after spanning confirmed
When inserting a long word into the grid, we don't actually confirm that
the grid has been spanned before breaking. It turns out that there's a
bug somewhere, somehow, in the route-choosing logic which makes it
possible for a grid to not be spanned.

When we detect this situation, instead of returning an unsolveable grid,
let's just throw everything away and start over with a new long-word
path.

The ideal situation would be to fix that bug, but .. I don't feel like
doing that now. The repro case for this is: `4x4/1737621998508`